requirements.txt
pyspark==3.1.2
pandas==1.4.3
jupyter==1.0.0
python-dateutil == 2.8.2
python-dotenv == 0.21.0
# google-cloud-bigquery==3.4.0
# pandas-gbq==0.18.1